+++ example/lib/main.dart
@@ -12,15 +12,15 @@ class MyApp extends StatefulWidget {
 }
 
 class _MyAppState extends State<MyApp> {
-  final twitterLogin =
-  new ShareSocialMediaPlugin(consumerKey: "3WGlyun7pWXYP6s5GjFiaCFCI", consumerSecret: 'pyNN593fU4hHOvSEcatcXAo1epk5pv1f2T6rAYMuXqyZgMH0OT');
+  final twitterLogin = new ShareSocialMediaPlugin(
+      consumerKey: "YOUR_CONSUMER_KEY",
+      consumerSecret: 'CONSUMER_SECRECT');
 
-   Future<int> _startSession() async {
+  Future<int> _startSession() async {
     var sessionData = await twitterLogin.currentSession;
 
     if (sessionData == null) {
       var result = await twitterLogin.authorize();
-      print(result.session);
     } else
       print(sessionData);
 
@@ -28,21 +28,16 @@ class _MyAppState extends State<MyApp> {
   }
 
   void _logout() async {
-    var twitterLogin =
-    new ShareSocialMediaPlugin(consumerKey: "3WGlyun7pWXYP6s5GjFiaCFCI", consumerSecret: 'pyNN593fU4hHOvSEcatcXAo1epk5pv1f2T6rAYMuXqyZgMH0OT');
     await twitterLogin.logOut();
-
   }
 
   @override
   void initState() {
     super.initState();
-
   }
 
   @override
   Widget build(BuildContext context) {
-
     return MaterialApp(
       home: Scaffold(
         appBar: AppBar(
@@ -54,15 +49,13 @@ class _MyAppState extends State<MyApp> {
               onPressed: () async {
                 await ShareSocialMediaPlugin.shareLine("http://www.google.com");
               },
-              child: Text('Line', style: TextStyle(fontSize: 20)),
-            ),
-            RaisedButton(
-              onPressed: () { _startSession();},
-              child: Text('Twitter', style: TextStyle(fontSize: 20)),
+              child: Text('Share in Line', style: TextStyle(fontSize: 20)),
             ),
             new RaisedButton(
-              child: new Text('Log out'),
-              onPressed: _logout,
+              child: new Text('Share in Twitter'),
+              onPressed: () {
+                twitterLogin.shareTwitter("hola mundo");
+              },
             ),
           ],
         ),